# docs/openapi/locker/openapi.yaml
openapi: "3.0.0"
info:
  description: Locker API
  title: Locker
  version: poc 0.0.1
paths:
...
  /v1/locks:
    post:
      operationId: create Lock
      description: |
        Create or obtain a lock for a collection of `KeyValues`. This request blocks if another process
        already has the lock being requested. It will eventually return if the lock has been freed and
        claimed, or the server is restarting. In which case, the client will need to retry claiming the lock
      requestBody:
        required: true
        content:
          appplication/json:
            schema:
              type: object
              required:
                - KeyValues
              properties:
                Timeout:
                  type: integer
                  format: int64
                  default: 0
                  description: "If set to 0, then the server's configuration will be used"
                KeyValues:
                  minLength: 1
                  $ref: "../common/components.yaml#/components/schemas/KeyValues"
      responses:
        200:
          description: created or obtained the lock
          content:
            appplication/json:
              schema:
                type: object
                properties:
                  SessionID:
                    type: string
                    description: "Session ID that is used to send hertbeats or release obtained locks"
                  Timeout:
                    description: "How long a server will keep the lock. This value is refreshed when a heartbeat is received"
                    type: integer
                    format: int64
        400:
          description: returned when an error processing for the request body fails
          content:
            appplication/json:
              schema:
                $ref: "../common/components.yaml#/components/schemas/ApiError"
        503:
          description: server is shutting down for a restart and client should retry obtaining the lock
    get:
      operationId: query Locks
      description: |
        Query all the locks currently in use by this service
      requestBody:
        required: true
        content:
          appplication/json:
            schema:
              $ref: "../common/components.yaml#/components/schemas/AssociatedKeyValuesQuery"
      responses:
        200:
          description: OK
          content:
            appplication/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
        400:
          description: returned when an error processing for the request body fails
          content:
            appplication/json:
              schema:
                $ref: "../common/components.yaml#/components/schemas/ApiError"
...
components:
  schemas:
...
    ListResponse:
      type: object
      properties:
        Locks:
          type: array
          items:
            type: object
            properties:
              SessionID:
                type: string
              KeyValues:
                $ref: "../common/components.yaml#/components/schemas/KeyValues"
              Timeout:
                type: integer
                format: int64
              TimeTillExpire:
                type: integer
                format: int64
              LocksHeldOrWaiting:
                type: integer